decomponet product list
