+++ src/blocks/ProductList/blocks/Filters/index.js
@@ -1,0 +1,1 @@
+export { Filters } from './Filters';
